+++ test/fields/collections/Collapsible/index.ts
@@ -133,19 +133,4 @@ const CollapsibleFields: CollectionConfig = {
   ],
 }
 
-export const collapsibleDoc = {
-  text: 'Seeded collapsible doc',
-  group: {
-    textWithinGroup: 'some text within a group',
-    subGroup: {
-      textWithinSubGroup: 'hello, get out',
-    },
-  },
-  arrayWithCollapsibles: [
-    {
-      innerCollapsible: '',
-    },
-  ],
-}
-
 export default CollapsibleFields