+++ test/fields/collections/Relationship/index.ts
@@ -15,6 +15,12 @@ const RelationshipFields: CollectionConfig = {
       type: 'relationship',
     },
     {
+      name: 'relationHasManyPolymorphic',
+      type: 'relationship',
+      relationTo: ['text-fields', 'array-fields'],
+      hasMany: true,
+    },
+    {
       name: 'relationToSelf',
       relationTo: relationshipFieldsSlug,
       type: 'relationship',
@@ -75,6 +81,13 @@ const RelationshipFields: CollectionConfig = {
         },
       ],
       type: 'array',
+    },
+    {
+      name: 'relationshipWithMinRows',
+      relationTo: ['text-fields'],
+      hasMany: true,
+      minRows: 2,
+      type: 'relationship',
     },
   ],
   slug: relationshipFieldsSlug,