+++ test/fields/collections/Text/BeforeInput.tsx
@@ -1,0 +1,9 @@
+'use client'
+
+import React from 'react'
+
+const BeforeInputImpl: React.FC = () => {
+  return <label className="before-input">#before-input</label>
+}
+
+export const BeforeInput = <BeforeInputImpl />