+++ test/fields/collections/Text/CustomLabel.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+import React from 'react'
+
+const CustomLabel: React.FC<{ htmlFor: string }> = ({ htmlFor }) => {
+  return (
+    <label htmlFor={htmlFor} className="custom-label">
+      #label
+    </label>
+  )
+}
+
+export default CustomLabel