+++ test/fields/collections/Upload/shared.ts
@@ -1,0 +1,5 @@
+import type { Upload } from '../../payload-types'
+
+export const uploadsDoc: Partial<Upload> = {
+  text: 'An upload here',
+}