+++ test/fields/e2e.spec.ts
@@ -6,14 +6,13 @@ import path from 'path'
 import payload from '../../packages/payload/src'
 import { mapAsync } from '../../packages/payload/src/utilities/mapAsync'
 import wait from '../../packages/payload/src/utilities/wait'
-import { saveDocAndAssert, saveDocHotkeyAndAssert } from '../helpers'
+import { exactText, saveDocAndAssert, saveDocHotkeyAndAssert } from '../helpers'
 import { AdminUrlUtil } from '../helpers/adminUrlUtil'
 import { initPayloadE2E } from '../helpers/configHelpers'
 import { RESTClient } from '../helpers/rest'
-import { jsonDoc } from './collections/JSON'
-import { numberDoc } from './collections/Number'
-import { textDoc } from './collections/Text'
-import { lexicalE2E } from './lexicalE2E'
+import { jsonDoc } from './collections/JSON/shared'
+import { numberDoc } from './collections/Number/shared'
+import { textDoc } from './collections/Text/shared'
 import { clearAndSeedEverything } from './seed'
 import {
   collapsibleFieldsSlug,
@@ -23,7 +22,7 @@ import {
   textFieldsSlug,
 } from './slugs'
 
-const { afterEach, beforeAll, describe, beforeEach } = test
+const { afterEach, beforeAll, beforeEach, describe } = test
 
 let client: RESTClient
 let page: Page
@@ -34,7 +33,7 @@ describe('fields', () => {
   beforeAll(async ({ browser }) => {
     const config = await initPayloadE2E(__dirname)
     serverURL = config.serverURL
-    client = new RESTClient(null, { serverURL, defaultSlug: 'users' })
+    client = new RESTClient(null, { defaultSlug: 'users', serverURL })
     await client.login()
 
     const context = await browser.newContext()
@@ -43,13 +42,13 @@ describe('fields', () => {
   beforeEach(async () => {
     await clearAndSeedEverything(payload)
     await client.logout()
-    client = new RESTClient(null, { serverURL, defaultSlug: 'users' })
+    client = new RESTClient(null, { defaultSlug: 'users', serverURL })
     await client.login()
   })
   describe('text', () => {
     let url: AdminUrlUtil
     beforeAll(() => {
-      url = new AdminUrlUtil(serverURL, 'text-fields')
+      url = new AdminUrlUtil(serverURL, textFieldsSlug)
     })
 
     test('should display field in list view', async () => {
@@ -79,6 +78,40 @@ describe('fields', () => {
       await page.goto(url.create)
       const description = page.locator('.field-description-i18nText')
       await expect(description).toHaveText('en description')
+    })
+
+    test('should render custom label', async () => {
+      await page.goto(url.create)
+      const label = page.locator('label.custom-label[for="field-customLabel"]')
+      await expect(label).toHaveText('#label')
+    })
+
+    test('should render custom error', async () => {
+      await page.goto(url.create)
+      const input = page.locator('input[id="field-customError"]')
+      await input.fill('ab')
+      await expect(input).toHaveValue('ab')
+      const error = page.locator('.custom-error:near(input[id="field-customError"])')
+      const submit = page.locator('button[type="button"][id="action-save"]')
+      await submit.click()
+      await expect(error).toHaveText('#custom-error')
+    })
+
+    test('should render BeforeInput and AfterInput', async () => {
+      await page.goto(url.create)
+      const input = page.locator('input[id="field-beforeAndAfterInput"]')
+
+      const prevSibling = await input.evaluateHandle((el) => {
+        return el.previousElementSibling
+      })
+      const prevSiblingText = await page.evaluate((el) => el.textContent, prevSibling)
+      await expect(prevSiblingText).toEqual('#before-input')
+
+      const nextSibling = await input.evaluateHandle((el) => {
+        return el.nextElementSibling
+      })
+      const nextSiblingText = await page.evaluate((el) => el.textContent, nextSibling)
+      await expect(nextSiblingText).toEqual('#after-input')
     })
   })
 
@@ -153,6 +186,25 @@ describe('fields', () => {
       await saveDocAndAssert(page)
       await expect(field.locator('.rs__value-container')).toContainText(String(input))
     })
+
+    test('should bypass min rows validation when no rows present and field is not required', async () => {
+      await page.goto(url.create)
+      await saveDocAndAssert(page)
+      await expect(page.locator('.Toastify')).toContainText('successfully')
+    })
+
+    test('should fail min rows validation when rows are present', async () => {
+      const input = 5
+
+      await page.goto(url.create)
+      await page.locator('.field-withMinRows').click()
+
+      await page.keyboard.type(String(input))
+      await page.keyboard.press('Enter')
+      await page.click('#action-save', { delay: 100 })
+
+      await expect(page.locator('.Toastify')).toContainText('Please correct invalid fields')
+    })
   })
 
   describe('indexed', () => {
@@ -161,16 +213,17 @@ describe('fields', () => {
       url = new AdminUrlUtil(serverURL, 'indexed-fields')
     })
 
+    // TODO: This test is flaky
     test('should display unique constraint error in ui', async () => {
       const uniqueText = 'uniqueText'
       await payload.create({
         collection: 'indexed-fields',
         data: {
-          text: 'text',
-          uniqueText,
           group: {
             unique: uniqueText,
           },
+          text: 'text',
+          uniqueText,
         },
       })
 
@@ -286,17 +339,17 @@ describe('fields', () => {
       filledGroupPoint = await payload.create({
         collection: pointFieldsSlug,
         data: {
-          point: [5, 5],
-          localized: [4, 2],
           group: { point: [4, 2] },
+          localized: [4, 2],
+          point: [5, 5],
         },
       })
       emptyGroupPoint = await payload.create({
         collection: pointFieldsSlug,
         data: {
-          point: [5, 5],
-          localized: [3, -2],
           group: {},
+          localized: [3, -2],
+          point: [5, 5],
         },
       })
     })
@@ -530,6 +583,38 @@ describe('fields', () => {
       ).toHaveValue('items>1>title')
     })
 
+    test('should bypass min rows validation when no rows present and field is not required', async () => {
+      await page.goto(url.create)
+      await saveDocAndAssert(page)
+      await expect(page.locator('.Toastify')).toContainText('successfully')
+    })
+
+    test('should fail min rows validation when rows are present', async () => {
+      await page.goto(url.create)
+
+      await page
+        .locator('#field-blocksWithMinRows')
+        .getByRole('button', { name: 'Add Blocks With Min Row' })
+        .click()
+
+      const blocksDrawer = page.locator('[id^=drawer_1_blocks-drawer-]')
+      await expect(blocksDrawer).toBeVisible()
+
+      const firstBlockSelector = blocksDrawer
+        .locator('.blocks-drawer__blocks .blocks-drawer__block')
+        .first()
+
+      await firstBlockSelector.click()
+
+      const firstRow = page.locator('input[name="blocksWithMinRows.0.blockTitle"]')
+      await expect(firstRow).toBeVisible()
+      await firstRow.fill('first row')
+      await expect(firstRow).toHaveValue('first row')
+
+      await page.click('#action-save', { delay: 100 })
+      await expect(page.locator('.Toastify')).toContainText('Please correct invalid fields')
+    })
+
     describe('row manipulation', () => {
       describe('react hooks', () => {
         test('should add 2 new block rows', async () => {
@@ -603,6 +688,20 @@ describe('fields', () => {
         '#rowLabelAsComponent-row-0 >> .row-label :text("custom row label")',
       )
       await expect(customRowLabel).toHaveCSS('text-transform', 'uppercase')
+    })
+
+    test('should bypass min rows validation when no rows present and field is not required', async () => {
+      await page.goto(url.create)
+      await saveDocAndAssert(page)
+      await expect(page.locator('.Toastify')).toContainText('successfully')
+    })
+
+    test('should fail min rows validation when rows are present', async () => {
+      await page.goto(url.create)
+      await page.locator('#field-arrayWithMinRows >> .array-field__add-row').click()
+
+      await page.click('#action-save', { delay: 100 })
+      await expect(page.locator('.Toastify')).toContainText('Please correct invalid fields')
     })
 
     describe('row manipulation', () => {
@@ -762,7 +861,6 @@ describe('fields', () => {
       )
     })
   })
-  describe('lexical', lexicalE2E(client, page, serverURL))
   describe('richText', () => {
     async function navigateToRichTextFields() {
       const url: AdminUrlUtil = new AdminUrlUtil(serverURL, 'rich-text-fields')
@@ -1139,8 +1237,8 @@ describe('fields', () => {
       describe('EST', () => {
         test.use({
           geolocation: {
-            longitude: -83.0458,
             latitude: 42.3314,
+            longitude: -83.0458,
           },
           timezoneId: 'America/Detroit',
         })
@@ -1160,7 +1258,7 @@ describe('fields', () => {
           const id = routeSegments.pop()
 
           // fetch the doc (need the date string from the DB)
-          const { doc } = await client.findByID({ id, slug: 'date-fields', auth: true })
+          const { doc } = await client.findByID({ id, auth: true, slug: 'date-fields' })
 
           expect(doc.default).toEqual('2023-02-07T12:00:00.000Z')
         })
@@ -1169,8 +1267,8 @@ describe('fields', () => {
       describe('PST', () => {
         test.use({
           geolocation: {
-            longitude: -122.419416,
             latitude: 37.774929,
+            longitude: -122.419416,
           },
           timezoneId: 'America/Los_Angeles',
         })
@@ -1191,7 +1289,7 @@ describe('fields', () => {
           const id = routeSegments.pop()
 
           // fetch the doc (need the date string from the DB)
-          const { doc } = await client.findByID({ id, slug: 'date-fields', auth: true })
+          const { doc } = await client.findByID({ id, auth: true, slug: 'date-fields' })
 
           expect(doc.default).toEqual('2023-02-07T12:00:00.000Z')
         })
@@ -1200,8 +1298,8 @@ describe('fields', () => {
       describe('ST', () => {
         test.use({
           geolocation: {
-            longitude: -171.857,
             latitude: -14.5994,
+            longitude: -171.857,
           },
           timezoneId: 'Pacific/Apia',
         })
@@ -1222,7 +1320,7 @@ describe('fields', () => {
           const id = routeSegments.pop()
 
           // fetch the doc (need the date string from the DB)
-          const { doc } = await client.findByID({ id, slug: 'date-fields', auth: true })
+          const { doc } = await client.findByID({ id, auth: true, slug: 'date-fields' })
 
           expect(doc.default).toEqual('2023-02-07T12:00:00.000Z')
         })
@@ -1245,7 +1343,7 @@ describe('fields', () => {
       })
       const relationshipIDs = allRelationshipDocs.docs.map((doc) => doc.id)
       await mapAsync(relationshipIDs, async (id) => {
-        await payload.delete({ collection: relationshipFieldsSlug, id })
+        await payload.delete({ id, collection: relationshipFieldsSlug })
       })
     })
 
@@ -1352,6 +1450,41 @@ describe('fields', () => {
       await page.locator('.rs__option:has-text("Seeded text document")').click()
       await field.locator('.clear-indicator').click()
       await expect(field.locator('.rs__placeholder')).toBeVisible()
+    })
+
+    test('should display `hasMany` polymorphic relationships', async () => {
+      await page.goto(url.create)
+      const field = page.locator('#field-relationHasManyPolymorphic')
+      await field.click()
+
+      await page
+        .locator('.rs__option', {
+          hasText: exactText('Seeded text document'),
+        })
+        .click()
+
+      await expect(
+        page
+          .locator('#field-relationHasManyPolymorphic .relationship--multi-value-label__text', {
+            hasText: exactText('Seeded text document'),
+          })
+          .first(),
+      ).toBeVisible()
+
+      // await fill the required fields then save the document and check again
+      await page.locator('#field-relationship').click()
+      await page.locator('#field-relationship .rs__option:has-text("Seeded text document")').click()
+      await saveDocAndAssert(page)
+
+      const valueAfterSave = page.locator('#field-relationHasManyPolymorphic .multi-value').first()
+
+      await expect(
+        valueAfterSave
+          .locator('.relationship--multi-value-label__text', {
+            hasText: exactText('Seeded text document'),
+          })
+          .first(),
+      ).toBeVisible()
     })
 
     test('should populate relationship dynamic default value', async () => {
@@ -1476,6 +1609,34 @@ describe('fields', () => {
       expect(seededTextDocument.docs.length).toEqual(1)
       // but the relationship document should NOT exist, as the hotkey should have saved the drawer and not the parent page
       expect(relationshipDocuments.docs.length).toEqual(0)
+    })
+
+    test('should bypass min rows validation when no rows present and field is not required', async () => {
+      await page.goto(url.create)
+      // First fill out the relationship field, as it's required
+      await page.locator('#relationship-add-new .relationship-add-new__add-button').click()
+      await page.locator('#field-relationship .value-container').click()
+      await page.getByText('Seeded text document', { exact: true }).click()
+
+      await saveDocAndAssert(page)
+      await expect(page.locator('.Toastify')).toContainText('successfully')
+    })
+
+    test('should fail min rows validation when rows are present', async () => {
+      await page.goto(url.create)
+
+      // First fill out the relationship field, as it's required
+      await page.locator('#relationship-add-new .relationship-add-new__add-button').click()
+      await page.locator('#field-relationship .value-container').click()
+      await page.getByText('Seeded text document', { exact: true }).click()
+
+      await page.locator('#field-relationshipWithMinRows .value-container').click()
+      await page
+        .locator('#field-relationshipWithMinRows .rs__option:has-text("Seeded text document")')
+        .click()
+
+      await page.click('#action-save', { delay: 100 })
+      await expect(page.locator('.Toastify')).toContainText('Please correct invalid fields')
     })
   })